==> 1-process.ipynb <==
# **Data Processing**

_John Andrew Dixon_

---

##### **Imports**

# Imports 
import pandas as pd
import numpy as np

##### **Data Load**

# Remote URL to the IMDB data
basics_url = "https://datasets.imdbws.com/title.basics.tsv.gz"
# Loading the remote data
basics_df = pd.read_csv(basics_url, sep="\t", low_memory=False)
# Verify
basics_df

# Remote URL to the IMDB data
akas_url = "https://datasets.imdbws.com/title.akas.tsv.gz"
# Loading the remote data
akas_df = pd.read_csv(akas_url, sep="\t", low_memory=False)
# Verify
akas_df

# Remote URL to the IMDB data
ratings_url = "https://datasets.imdbws.com/title.ratings.tsv.gz"
# Loading the remote data
ratings_df = pd.read_csv(ratings_url, sep="\t", low_memory=False)
# Verify
ratings_df

---

## **Processing**

##### _AKAs_

# Replace "\N" with np.nan
akas_df.replace({"\\N": np.nan}, inplace=True)

##### _Ratings_

# Replace "\N" with np.nan
ratings_df.replace({"\\N": np.nan}, inplace=True)

##### _Basics_

# Replace "\N" with np.nan
basics_df.replace({"\\N": np.nan}, inplace=True)
basics_df["startYear"] = basics_df["startYear"].astype(float)
basics_df["endYear"] = basics_df["endYear"].astype(float)

# Eliminate movies that are null for runtimeMinutes
basics_df.dropna(axis=0, subset=["runtimeMinutes"], inplace=True)
basics_df

# Eliminate movies that are null for genre
basics_df.dropna(axis=0, subset=["genres"], inplace=True)
basics_df

# Keep only titleType==Movie

# Create a filter than gets titleType = Movie for all rows
is_movie = basics_df["titleType"] == "movie"

# Apply the filter
basics_df = basics_df.loc[is_movie, :]
basics_df

# Create a filter that selects all movies with a startYear between 2000 and 2022
between_2000_2022 = (basics_df["startYear"] >= 2000) & (basics_df["startYear"] <= 2022)

# Apply the filter
basics_df = basics_df.loc[between_2000_2022, :]
basics_df

# Eliminate movies that include "Documentary" in genre

# Create a filter that gets all non-documentary movies
is_not_documentary = ~basics_df["genres"].str.contains("documentary", case=False)

# Apply the filter
basics_df = basics_df.loc[is_not_documentary, :]
basics_df

##### _Keep Only US Movies_

# Keep only US movies.
akas_df = akas_df.loc[akas_df["region"] == "US", :]
akas_df

# Basics: Keep only US movies

# Get the ratings that have tconst within the basics titleId
only_US = basics_df["tconst"].isin(akas_df["titleId"])

# Apply the filter
basics_df = basics_df.loc[only_US, :]
basics_df

# Keep only US movies

# Get the ratings that have tconst within the akas titleId
only_US = ratings_df["tconst"].isin(akas_df["titleId"])

# Apply the filter
ratings_df = ratings_df.loc[only_US, :]
ratings_df

## **Pre-save check**

_Before saving, run a final .info() for each of the dataframes to show a summary of how many movies remain and the datatypes of each feature._

basics_df.info()

akas_df.info()

ratings_df.info()

## **Saving Results**

# Save each DataFrame to compressed files
basics_df.to_csv("Data/title_basics.csv.gz", compression='gzip',index=False)
ratings_df.to_csv("Data/title_ratings.csv.gz", compression='gzip',index=False)
akas_df.to_csv("Data/title_akas.csv.gz", compression='gzip',index=False)

# Open saved file and preview again
basics_df = pd.read_csv("Data/title_basics.csv.gz", low_memory=False)
basics_df

# Open saved file and preview again
ratings_df = pd.read_csv("Data/title_ratings.csv.gz", low_memory=False)
ratings_df

# Open saved file and preview again
akas_df = pd.read_csv("Data/title_akas.csv.gz", low_memory=False)
akas_df
