==> 2a-efficient-API.ipynb <==
# **TMBD API Extraction**

_John Andrew Dixon_

---

##### **Imports**

import json, os, time
import pandas as pd
import tmdbsimple as tmdb
from tqdm.notebook import tqdm_notebook

##### **Helper Functions**

def get_movie_with_rating(movie_id):
    """Using the TMDB API, retrieves movie info and rating based on a given movie_id"""

    # Using the TMDB API, get movie based on "movie_id"
    movie = tmdb.Movies(movie_id)

    # Save movie info to a variable
    info = movie.info()
    # Save movie releases to a variable
    releases = movie.releases()

    # Loop through all releases for each country
    for country in releases["countries"]:
        # If the country is US, add the certification to the info dictionary
        if country["iso_3166_1"] == "US":
            info["certification"] = country["certification"]

    # Return the info dictionary
    return info

def write_json(new_data, filename):
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""

    with open(filename, "r+") as file:
        # Load existing data as a dictionary
        file_data = json.load(file)

        # Choose to extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
            file_data.append(new_data)

        # Set file's current position at offset
        file.seek(0)

        # Convert back to JSON
        json.dump(file_data, file)

##### **Data Load**

# Load the previous "basics" file from part 1
basics_df = pd.read_csv("Data/title_basics.csv.gz")
# Verify it loaded
basics_df.head()

---

## **Test API Function**

# Load API Key
with open("/Users/johna/.secret/tmdb_api.json", "r") as f:
    login = json.load(f)
login.keys()

# Link API Key with API
tmdb.API_KEY = login["api-key"]

# IDs to test. The first ID = "Avengers"; the last ID = "The Notebook"
# The middle id = error test for a non-existent movie
test_ids = ["tt0848228", "tt0115937","tt0332280"]
# List to hold the results
results = []
# List to hold errors given by the API
errors = []

# Iterate through the test IDs
for movie_id in test_ids:
    # Try to get the current test_id's movie with the get_movie_with_rating function
    try:
        # Save the result of the function into movie_info
        movie_info = get_movie_with_rating(movie_id)
        # Append result to the results list
        results.append(movie_info)
        
    # Handle errors if movie is not found
    except Exception as e:
        # Append error to errors list
        errors.append((movie_id, e))
    
pd.DataFrame(results)

# Display the errors
errors

This works, so let's move forward.

## **Efficient TMBD API Extraction**

# Specify the folder where data will be saved
FOLDER = "Data/"
# Create the folder; if it exists, move on
os.makedirs(FOLDER, exist_ok=True)
# List folder's contents
os.listdir(FOLDER)

# Specify years to get
YEARS_TO_GET = list(range(2002, 2011))
# Create empty list to hold potential errors
errors = []

############################## OUTER LOOP ############################## 
# Loop through the years and display progress as progress bar
for YEAR in tqdm_notebook(YEARS_TO_GET, desc="Years", position=0):
    # Create the path to the JSON file for the current YEAR in FOLDER
    JSON_FILE = f"{FOLDER}tmdb_api_results_{YEAR}.json"

    # Check if JSON_FILE exists
    file_exists = os.path.isfile(JSON_FILE)
    # If the file does not exists, create in the OS and populate it
    if not file_exists:
        with open(JSON_FILE, "w") as file:
            json.dump([{"imdb_id": 0}], file)

    # Filter the "basics_df" by the current YEAR
    current_year_df = basics_df.loc[basics_df["startYear"] == YEAR].copy()
    # Get the movie IDs from movies of the current year
    movie_ids = current_year_df["tconst"].copy()

    # Load in previous results
    previous_results_df = pd.read_json(JSON_FILE)
    # Filter out the movie ids to get based on what was previous gotten
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_results_df["imdb_id"])]

    ############################## INNER LOOP ############################## 
    # Inner loop to iterate through the movie we need per year of the outer loop
    for movie_id in tqdm_notebook(movie_ids_to_get, desc=f"Movies from {YEAR}", position=1, leave=True):
        try:
            # Get movie data for the current movie_id
            current_movie = get_movie_with_rating(movie_id)

            # Write the movie data to the proper file based on its year
            write_json(current_movie, JSON_FILE)

            # Wait 20 ms to prevent API overload
            time.sleep(0.02)
        # If there is an error, handle it
        except Exception as error:
            # Append all errors to the errors list
            errors.append([movie_id, error])
    
    # Read in the current year's API results from the file written to in the inner loop
    final_year_df = pd.read_json(JSON_FILE)
    # Compress the JSON file just created into a smaller format
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

# Output the errors from the results
print(f"[i] Total errors: {len(errors)}")
