==> 3-SQL.ipynb <==
# **SQL Database**

_John Andrew Dixon_

---

##### **Imports**

import pandas as pd
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import create_database, database_exists

##### **Data Load**

# Make it so that all of the dataframe's columns are shown
pd.set_option('display.max_columns', None)

title_basics_df = pd.read_csv("Data/title_basics.csv.gz")
title_basics_df.head()

title_ratings_df =  pd.read_csv("Data/title_ratings.csv.gz")
title_ratings_df.head()

TMDb_API_df = pd.read_csv("Data/first_decade_combined.csv.gz")
TMDb_API_df.info()

---

## **Database Creation**

# Format the connection to make
connection = "mysql+pymysql://root:root@localhost/movies"

# Create the connection engine
engine = create_engine(connection)

# Check if the database exists
if database_exists(connection):
    # Notify that the database exists
    print(f"The movies database exists!")
else:
    create_database(connection)

## **Table Creation**

##### **_title_ratings_**

# View Datatype info about the ratings DataFrame
title_ratings_df.dtypes

# Get the max length of the primary key "tconst"
key_len = title_ratings_df["tconst"].fillna("").map(len).max()

# Create the ratings schema
ratings_schema = {
    "tconst": String(key_len + 1),
    "averageRating": Float(),
    "numVotes": Integer()
}
ratings_schema

# Actually create the ratings table in the database
title_ratings_df.to_sql("title_ratings",
                  engine,
                  dtype=ratings_schema,
                  if_exists="replace",
                  index=False)

# Set "tconst" as the primary key
engine.execute("ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);");

# View the schema to verify the primary key was added
query = """DESCRIBE title_ratings;"""
pd.read_sql(query, engine)

# See the first 5 values of the "title_ratings" table
query = """
SELECT * FROM title_ratings
LIMIT 5;
"""
pd.read_sql(query, engine)

##### **_title_genres_**

# Split genres in the basics dataframe
title_basics_df["genres_split"] = title_basics_df["genres"].str.split(",")
# Explode the genres and get the unique values
unique_genres = sorted(title_basics_df["genres_split"].explode().unique())
unique_genres

# Explode the "genres_split" column
title_genres_df = title_basics_df.explode("genres_split")
# Get all the rows but only the "genres_split" and "tconst"
title_genres_df = title_genres_df[["tconst", "genres_split"]]
title_genres_df.head()

# Rename the column to reflect the futue genre_id that'll replace the genre name
title_genres_df.rename(columns={"genres_split": "genre_id"}, inplace=True)
title_genres_df.head()

# Create a genre map
genre_map = dict(zip(unique_genres, range(len(unique_genres))))
genre_map

title_genres_df["genre_id"] = title_genres_df["genre_id"].map(genre_map)
title_genres_df.head()

# Actually create the ratings table in the database
title_genres_df.to_sql("title_genres",
                        engine,
                        if_exists="replace",
                        index=False)

# View the schema to verify the primary key was added
query = """DESCRIBE title_genres;"""
pd.read_sql(query, engine)

# See the first 5 values of the "title_genres" table
query = """
SELECT * FROM title_genres
LIMIT 5;
"""
pd.read_sql(query, engine)

##### _genres_

# Make the genre map into a dataframe
genres_df = pd.DataFrame({
    "genre_id": genre_map.values(),
    "genre_name": genre_map.keys()
})

# Add the table to the database
genres_df.to_sql("genres",
                 engine,
                 if_exists="replace",
                 index=False)

# Set "genre_id" as the primary key
engine.execute("ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);");

# View the schema to verify the primary key was added
query = """DESCRIBE genres;"""
pd.read_sql(query, engine)

# See the first 5 values of the "genres" table
query = """
SELECT * FROM genres
LIMIT 5;
"""
pd.read_sql(query, engine)

##### _tmdb_data_

# Extract only the relevant columns
TMDb_API_df = TMDb_API_df[["imdb_id", "revenue", "budget", "certification"]]
TMDb_API_df.info()

# Get the datatypes of each column
TMDb_API_df.dtypes

# Get the max length of the primary key "tconst"
key_len = TMDb_API_df["imdb_id"].fillna("").map(len).max()
# Get the max length of the certification key column "certification"
cert_len = TMDb_API_df["certification"].fillna("").map(len).max()

# Create the TMDB_data schema
TMDb_data_schema = {
    "imdb_id": String(key_len + 1),
    "revenue": Float(),
    "budget": Float(),
    "certification": Text(cert_len + 1)
}
TMDb_data_schema

# Add the "tmdb_data" table to the database
TMDb_API_df.to_sql("tmdb_data",
                   engine,
                   dtype=TMDb_data_schema,
                   if_exists="replace",
                   index=False)

# Set "imdb_id" as the primary key
engine.execute("ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);");

# View the schema to verify the primary key was added
query = """DESCRIBE tmdb_data;"""
pd.read_sql(query, engine)

# See the first 5 values of the "tmdb_data" table
query = """
SELECT * FROM tmdb_data
LIMIT 5;
"""
pd.read_sql(query, engine)

#### _title_basics_

# Extract the necessary info from "title_basics"
title_basics_df = title_basics_df[["tconst", "primaryTitle", "startYear", "endYear", "runtimeMinutes"]]
title_basics_df.rename(columns={
    "primaryTitle": "primary_title", 
    "startYear": "start_year", 
    "endYear": "end_year", 
    "runtimeMinutes": "runtime_minutes"
}, inplace=True)
title_basics_df.head()

# Get the datatypes of each column
title_basics_df.dtypes

# Get the max length of the primary key "tconst"
key_len = title_basics_df["tconst"].fillna("").map(len).max()
# Get the max length of the certification column "primary_title"
primary_title_len = title_basics_df["primary_title"].fillna("").map(len).max()

# Create the "title_basics" schema
title_basics_schema = {
    "tconst": String(key_len + 1),
    "primary_title": Text(primary_title_len + 1),
    "start_year": Float(),
    "end_year": Float(),
    "runtime_minutes": Integer()
}
title_basics_schema

# Add the "title_basics" table to the database
title_basics_df.to_sql("title_basics",
                       engine,
                       dtype=title_basics_schema,
                       if_exists="replace",
                       index=False)

# Set "imdb_id" as the primary key
engine.execute("ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);");

# View the schema to verify the primary key was added
query = """DESCRIBE title_basics;"""
pd.read_sql(query, engine)

# See the first 5 values of the "title_basics" table
query = """
SELECT * FROM title_basics
LIMIT 5;
"""
pd.read_sql(query, engine)

##### **_Final Table Verification_**

# Verify the tables were added by showing them
query = """SHOW TABLES;"""
pd.read_sql(query, engine)
